requirements.txt
flask>=2.0
requests>=2.28
sentence-transformers>=2.2.2
scikit-learn>=1.2
joblib>=1.2
python-dotenv>=1.0
gunicorn>=20.1.0
